### 2020-12-17 #30 @NickleDave
Hi @JonnyTran thank you for making this presubmission inquiry. Sorry we haven't gotten back to you sooner.

`OpenOmics` definitely looks like it is within scope for PyOpenSci, and I can see you have done a lot of work already to make sure the package adheres to our standards and guidelines, including your use of the pyopensci cookiecutter.

I am not sure if you saw that we're "on pause" for a bit while looking for further funding:
https://pyopensci.discourse.group/t/on-pause-for-a-bit/224

In the meantime, I have been trying to help where I can.
If @lwasser is okay with it, I could act as editor for this submission.

But realistically I won't be able to start that process until mid-January because of other obligations.
Also, I want to make sure you're aware that currently our total total time for review has been 1-2 months.

I definitely want to do what I can to recognize all the hard work you've done on `OpenOmics`, and help you further develop the library via the PyOpenSci review process.
But I also want to make sure we're on the same page about what the review process will look like because of those other factors.
Please let us know if those timelines would be acceptable to you.
