### 2020-12-17 #30 @JonnyTran
> But realistically I won't be able to start that process until mid-January because of other obligations.
> Also, I want to make sure you're aware that currently our total total time for review has been 1-2 months.
> Please let us know if those timelines would be acceptable to you.

Yes, I'm more than okay with the timeline. Thank you for the consideration of reviewing this work.
