+++ README.md
@@ -9,7 +9,7 @@
 This Python package provide a series of tool to integrate and query the genomics, transcriptomics, proteomics, and clinical TCGA data.
 By providing a series of data manipulation tools, open-omics facilitates the common coding tasks when preparing data for bioinformatics analysis.
 
-## Installation via pip:
+## What is OpenOmics?:
 OpenOmics is a Python library to assist integration of heterogeneous multi-omics bioinformatics data. By providing an API of data manipulation tools as well as a web interface, OpenOmics facilitates the common coding tasks when preparing data for bioinformatics analysis.
 It features support for:
 - Genomics, Transcriptomics, Proteomics, and Clinical data.
