### 2020-12-20 0b30547
refactors gene_index parameters
