+++ tests/test_multiomics.py
@@ -10,33 +10,35 @@ from openomics import MultiOmics
 
 cohort_folder_path = "tests/data/TCGA_LUAD"
 
+
 @pytest.fixture
 def generate_TCGA_LUAD_MessengerRNA():
-    return MessengerRNA("LUAD", data=os.path.join(cohort_folder_path, "LUAD__geneExp.txt"), transposed=True,
-                        columns="GeneSymbol|TCGA", gene_index_by="gene_name")
+    return MessengerRNA(data=os.path.join(cohort_folder_path, "LUAD__geneExp.txt"), transpose=True,
+                        usecols="GeneSymbol|TCGA", gene_index="gene_name")
+
 
 @pytest.fixture
 def generate_TCGA_LUAD_MicroRNA():
-    return MicroRNA("LUAD", data=os.path.join(cohort_folder_path, "LUAD__miRNAExp__RPM.txt"), transposed=True,
-                    columns="GeneSymbol|TCGA", gene_index_by="gene_name")
+    return MicroRNA(data=os.path.join(cohort_folder_path, "LUAD__miRNAExp__RPM.txt"), transpose=True,
+                    usecols="GeneSymbol|TCGA", gene_index="gene_name")
 
 
 @pytest.fixture
 def generate_TCGA_LUAD_LncRNA():
-    return LncRNA("LUAD", data=os.path.join(cohort_folder_path, "TCGA-rnaexpr.tsv"), transposed=True,
-                  columns="Gene_ID|TCGA", gene_index_by="gene_id")
+    return LncRNA(data=os.path.join(cohort_folder_path, "TCGA-rnaexpr.tsv"), transpose=True,
+                  usecols="Gene_ID|TCGA", gene_index="gene_id")
 
 
 @pytest.fixture
 def generate_TCGA_LUAD_SomaticMutation():
-    return SomaticMutation("LUAD", data=os.path.join(cohort_folder_path, "LUAD__somaticMutation_geneLevel.txt"),
-                           transposed=True, columns="GeneSymbol|TCGA", gene_index_by="gene_name")
+    return SomaticMutation(data=os.path.join(cohort_folder_path, "LUAD__somaticMutation_geneLevel.txt"),
+                           transpose=True, usecols="GeneSymbol|TCGA", gene_index="gene_name")
 
 
 @pytest.fixture
 def generate_TCGA_LUAD_Protein():
-    return Protein("LUAD", data=os.path.join(cohort_folder_path, "protein_RPPA.txt"), transposed=True,
-                   columns="GeneSymbol|TCGA", gene_index_by="protein_name")
+    return Protein(data=os.path.join(cohort_folder_path, "protein_RPPA.txt"), transpose=True,
+                   usecols="GeneSymbol|TCGA", gene_index="protein_name")
 
 
 def test_import_expression_table_size(generate_TCGA_LUAD_MessengerRNA):
