+++ openomics/transcriptomics.py
@@ -13,27 +13,30 @@ from .utils.df import drop_duplicate_columns
 
 
 class ExpressionData(object):
-    def __init__(self, cohort_name, data, transposed, columns=None, gene_index_by=None, sample_index_by="sample_index",
-                 log2_transform=False, dropna=False, npartitions=None):
+    def __init__(self, data, transpose, gene_index=None, usecols=None, gene_level=None, sample_level="sample_index",
+                 transform_fn=None, dropna=False, npartitions=None, cohort_name=None):
         """
         .. class:: ExpressionData
-        An abstract class that handles importing of any quantitative -omics data that is in a table format (e.g. csv, tsv, excel). Pandas will load the DataFrame from file with the user-specified columns and genes column name, then tranpose it such that the rows are samples and columns are gene/transcript/peptides.
+        This class handles importing of any quantitative omics data that is in a table format (e.g. csv, tsv, excel). Pandas will load the DataFrame from file with the user-specified columns and genes column name, then tranpose it such that the rows are samples and columns are gene/transcript/peptides.
         The user will also specify the index argument, which specifies if the genes are ensembl genes ID or gene name, or transcripts id/names. The user should be careful about choosing the right genes index which makes it easier to annotate functional, sequence, and interaction data to it.
         The dataframe should only contain numeric values besides the genes_col_name and the sample barcode id indices.
         Args:
+            gene_level:
             cohort_name (str): the unique cohort code name string
             data (str, byte-like, pandas.DataFrame):
                 Path or file stream of the table file to import. If a pandas DataFrame is passed, then import this dataframe and skip preprocessing steps.
-            transposed (bool): default True
+            transpose (bool):
                 True if sample names are columns and rows are genes. False if the table has samples for row index, and gene names as columns.
-            columns (str): a regex string
+            gene_index (str):
+                The column name of gene/transcript/protein to index by.
+            usecols (str): a regex string
                 A regex string to import column names from the table. Columns names imported are string match, separated by "|".
-            gene_index_by (str): {"gene_id", "transcript_id", "peptide_id", "gene_name", "trascript_name", "peptide_name"}
+            gene_level (str): {"gene", "transcript", "peptide"}
                 Chooses the level of the gene/transcript/peptide of the genes list in this expression data. The expression DataFrame's index will be renamed to this.
-            sample_index_by (str): {"sample_index", "patient_index"}
+            sample_level (str): {"sample_index", "patient_index"}
                 Chooses the level of the patient/sample/aliquot indexing.
-            log2_transform (bool): default False
-                Whether to log2 transform the expression values
+            transform_fn (bool): default False
+                A callable function to transform single values.
             dropna (bool):
                 Whether to drop rows with null values
             npartitions (int): [0-n], default 0
@@ -41,28 +44,33 @@ class ExpressionData(object):
         """
         self.cohort_name = cohort_name
 
-        df = self.load_dataframe(data, transposed=transposed, columns=columns, genes_index=gene_index_by)
-        self.expressions = self.preprocess_table(df, columns=columns, genes_index=gene_index_by, transposed=transposed,
+        df = self.load_dataframe(data, transpose=transpose, usecols=usecols, gene_index=gene_index)
+        self.expressions = self.preprocess_table(df, usecols=usecols, gene_index=gene_index, transposed=transpose,
                                                  dropna=dropna)
-        if npartitions:
+
+        # TODO load DD from file directly
[…]
