### 2020-12-22 #30 @NickleDave
> @NickleDave if you are keen to serve as an editor here i can work on the other submission! i've been overwhelmed with conference things but havea bit more space now to support.

Thank you @lwasser -- you mean Physcraper? I think I can handle this one and finish up that review as well.

> Yes, I'm more than okay with the timeline. Thank you for the consideration of reviewing this work.

Great glad to hear it @JonnyTran
Happy we can provide you a review and I'm looking forward to it.
I will be able to start the week of Jan 11, got it in my calendar now

Please, whenever you are ready, go ahead and open a new issue that's a full submission review.
If you can think of reviewers that would be appropriate, please also feel free to suggest them in the initial comment.
I will also look for reviewers with relevant expertise when I start the process the week of Jan. 11.
I can close this issue once you create the issue for full review.
Thank you!!!
