+++ README.md
@@ -26,8 +26,7 @@ OpenOmics also has an efficient data pipeline that bridges the popular data mani
 
     pip install openomics
 
-
-
+\n
 # How to use OpenOmics:
 
 
