### 2020-12-30 a3e6dd9
refactor database.df to .data
