### 2020-12-30 94766c3
added npartitions to read_gt
