+++ openomics/database/annotation.py
@@ -94,8 +94,8 @@ class RNAcentral(Dataset):
                            'external id': 'transcript_id',
                            'GO terms': 'go_id'}
 
-    def __init__(self, path="ftp://ftp.ebi.ac.uk/pub/databases/RNAcentral/current_release/",
-                 file_resources=None, col_rename=COLUMNS_RENAME_DICT, npartitions=0, species=9606, verbose=False):
+    def __init__(self, path="ftp://ftp.ebi.ac.uk/pub/databases/RNAcentral/current_release/", file_resources=None,
+                 col_rename=COLUMNS_RENAME_DICT, species=9606, npartitions=0, verbose=False):
         self.species = species
 
         if file_resources is None:
@@ -115,12 +115,22 @@ class RNAcentral(Dataset):
         gene_ids = []
         for file in file_resources:
             if "database_mappings" in file:
-                id_mapping = pd.read_table(file_resources[file],
-                                           low_memory=True, header=None,
-                                           names=["RNAcentral id", "database", "external id", "species", "RNA type",
-                                                  "gene symbol"])
+                if npartitions:
+                    id_mapping = dd.read_table(file_resources[file], header=None,
+                                               names=["RNAcentral id", "database", "external id", "species", "RNA type",
+                                                      "gene symbol"])
+                else:
+                    id_mapping = pd.read_table(file_resources[file],
+                                               low_memory=True, header=None,
+                                               names=["RNAcentral id", "database", "external id", "species", "RNA type",
+                                                      "gene symbol"])
+
                 gene_ids.append(id_mapping)
-        gene_ids = pd.concat(gene_ids, join="inner")
+
+        if npartitions:
+            gene_ids = dd.concat(gene_ids, join="inner")
+        else:
+            gene_ids = pd.concat(gene_ids, join="inner")
 
         gene_ids["species"] = gene_ids["species"].astype("O")
         if self.species is not None:
@@ -214,7 +224,10 @@ class NONCODE(Dataset):
         transcript2gene_df = pd.read_table(file_resources["NONCODEv5_Transcript2Gene"], header=None)
         transcript2gene_df.columns = ["NONCODE Transcript ID", "NONCODE Gene ID"]
 
-        self.noncode_func_df = pd.read_table(file_resources["NONCODEv5_human.func"], header=None)
+        if npartitions:
+            self.noncode_func_df = dd.read_table(file_resources["NONCODEv5_human.func"], header=None)
+        else:
+            self.noncode_func_df = pd.read_table(file_resources["NONCODEv5_human.func"], header=None)
         self.noncode_func_df.columns = ["NONCODE Gene ID", "GO terms"]
         self.noncode_func_df.set_index("NONCODE Gene ID", inplace=True)
 
@@ -296,24 +309,23 @@ class EnsemblGenes(BioMartManager, Dataset):
                           'gene_biotype', 'transcript_biotype', ]
         self.filename = "{}.{}".format(biomart, self.__class__.__name__)
         self.host = host
-        self.df = self.load_data(dataset=biomart, attributes=attributes, host=self.host,
-                                 filename=self.filename, npartitions=npartitions)
+        self.data = self.load_data(dataset=biomart, attributes=attributes, host=self.host,
+                                   filename=self.filename, npartitions=npartitions)
 
-        self.df.rename(columns=self.COLUMNS_RENAME_DICT,
-                       inplace=True)
-        print(self.name(), self.df.columns.tolist())
[…]
