+++ README.md
@@ -24,7 +24,7 @@ OpenOmics also has an efficient data pipeline that bridges the popular data mani
 
 ## Installation via pip: 
 
-> pip install openomics
+    pip install openomics
 
 <br/>
 
