+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.7.9
+current_version = 0.7.10
 commit = True
 tag = True
 
@@ -22,4 +22,3 @@ test = pytest
 
 [tool:pytest]
 collect_ignore = ['setup.py']
-
