+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.7.10
+current_version = v0.8.0
 commit = True
 tag = True
 
