+++ setup.py
@@ -52,7 +52,7 @@ with open('HISTORY.rst') as history_file:
 NAME = 'openomics'
 version = '0.8.0'
 DESCRIPTION = 'OpenOmics provides a bioinformatics API and web-app platform integrate and visualize the multiomics and clinical data.'
-URL = 'https://github.com/JonnyTran/OpenOmics'
+URL = 'https://github.com/BioMeCIS-Lab/OpenOmics'
 EMAIL = 'nhat.tran@mavs.uta.edu'
 
 AUTHOR = 'Jonny Tran'
