+++ docs/usage.rst
@@ -3,70 +3,76 @@ Usage
 =====
 
 To use openOmics in a project::
+::
 
-    import openomics
+   import openomics
 
-## Importing the openomics library & modules
+Importing the openomics library & modules
+-----------------------------------------
 
+.. code:: python
 
-```python
-from openomics import MultiOmics, MessengerRNA, MicroRNA, LncRNA
-```
+   from openomics import MultiOmics, MessengerRNA, MicroRNA, LncRNA
 
-## Import GTEx Tissue-specific gene expression dataset (directly from URL)
+Import GTEx Tissue-specific gene expression dataset (directly from URL)
+-----------------------------------------------------------------------
 
+.. code:: python
 
-```python
-from openomics.database import GTEx
+   from openomics.database import GTEx
 
-gtex = GTEx(path="https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/")
+   gtex = GTEx(path="https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/")
 
-gtex_transcripts = gtex.data
+   gtex_transcripts = gtex.data
 
-gene_id = pd.Index(gtex_transcripts["gene_id"].unique())
-gene_name = pd.Index(gtex_transcripts["gene_name"].unique())
-```
+   gene_id = pd.Index(gtex_transcripts["gene_id"].unique())
+   gene_name = pd.Index(gtex_transcripts["gene_name"].unique())
 
-## Import GENCODE human release 32
+Import GENCODE human release 32
+-------------------------------
 
-```python
-from openomics.database import GENCODE
+.. code:: python
 
-gencode = GENCODE(path="ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/",
-                  file_resources={"long_noncoding_RNAs.gtf": "gencode.v32.long_noncoding_RNAs.gtf.gz",
-                                  "basic.annotation.gtf": "gencode.v32.basic.annotation.gtf.gz",
-                                  "lncRNA_transcripts.fa": "gencode.v32.lncRNA_transcripts.fa.gz",
-                                  "transcripts.fa": "gencode.v32.transcripts.fa.gz"},
-                  remove_version_num=True)
-```
+   from openomics.database import GENCODE
 
-## Load single omics expressions for MessengerRNA, MicroRNA, LncRNA
-```python
-# Gene Expression
-messengerRNA_id = gtex_transcripts_gene_id & pd.Index(gencode.df[gencode.df["gene_type"] == "protein_coding"]["gene_id"].unique())
+   gencode = GENCODE(path="ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/",
+                     file_resources={"long_noncoding_RNAs.gtf": "gencode.v32.long_noncoding_RNAs.gtf.gz",
+                                     "basic.annotation.gtf": "gencode.v32.basic.annotation.gtf.gz",
+                                     "lncRNA_transcripts.fa": "gencode.v32.lncRNA_transcripts.fa.gz",
+                                     "transcripts.fa": "gencode.v32.transcripts.fa.gz"},
+                     remove_version_num=True)
 
-messengerRNA = MessengerRNA(gtex_transcripts[gtex_transcripts["gene_id"].isin(messengerRNA_id)],
-                            transpose=True, gene_index="gene_name", usecols=None, npartitions=4)
+Load single omics expressions for MessengerRNA, MicroRNA, LncRNA
[…]
