+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = v0.8.0
+current_version = 0.8.1
 commit = True
 tag = True
 
