+++ openomics/database/interaction.py
@@ -1,4 +1,5 @@
 from abc import abstractmethod
+import logging
 
 import networkx as nx
 from Bio import SeqIO
@@ -36,7 +37,7 @@ class Interactions(Dataset):
                 A dictionary to rename nodes in the network, where the nodes with name <dict[key]> will be renamed to <dict[value]>
         """
         # This class should NOT call super's __init__()
-        self.validate_file_resources(file_resources, path, verbose=verbose)
+        self.validate_file_resources(path, file_resources, verbose=verbose)
 
         self.data_path = path
         self.file_resources = file_resources
@@ -116,7 +117,7 @@ class Interactions(Dataset):
 
         for key, values in filters.items():
             if key not in df.columns:
-                print(f"Filter key `{key}` must be in one of {df.columns}")
+                logging.info(f"Filter key `{key}` must be in one of {df.columns}")
                 continue
             n_rows = df.shape[0]
 
@@ -130,7 +131,7 @@ class Interactions(Dataset):
             else:
                 df = df[df[key] == values]
 
-            print(f"INFO: Removed {n_rows - df.shape[0]} rows with `{key}` != {values}")
+            logging.info(f"INFO: Removed {n_rows - df.shape[0]} rows with `{key}` != {values}")
 
         assert df.shape[0] > 0, f"ERROR: Dataframe is empty because of filter: {filters}"
         return df
@@ -195,7 +196,7 @@ class BioGRID(Interactions):
                            #          'Modification', 'Phenotypes'],
                            low_memory=True)
 
-        print("{}: {}".format(self.name(), df.columns.tolist()))
+        logging.info("{}: {}".format(self.name(), df.columns.tolist()))
 
         df = self.filter_values(df, filters)
         network = nx.from_pandas_edgelist(df, source=source_col_name, target=target_col_name,
@@ -249,7 +250,7 @@ class STRING(Interactions, SequenceDataset):
     def load_network(self, file_resources, source_col_name, target_col_name, edge_attr, directed, filters):
         # protein_interactions = pd.read_table(file_resources["protein.links.txt"], sep=" ", low_memory=True)
         protein_interactions = pd.read_table(file_resources["protein.actions.txt"], sep="\t", low_memory=True)
-        print("{}: {}".format(self.name(), protein_interactions.columns.tolist()))
+        logging.info("{}: {}".format(self.name(), protein_interactions.columns.tolist()))
         protein_info = pd.read_table(file_resources["protein.info.txt"])
 
         self.protein_id2name = protein_info.set_index("protein_external_id")["preferred_name"].to_dict()
@@ -279,7 +280,7 @@ class STRING(Interactions, SequenceDataset):
 
             self.seq_dict[key] = sequence_str
 
-        print("Seq {} collisions: {}".format(index, collisions))
+        logging.info("Seq {} collisions: {}".format(index, collisions))
         return self.seq_dict
 
 
@@ -315,7 +316,7 @@ class LncBase(Interactions, Dataset):
         if edge_attr is None:
             edge_attr = ["tissue", "positive_negative"]
         df = pd.read_table(file_resources["LncBasev2_download.csv"], low_memory=True)
-        print(self.name(), df.columns.tolist())
+        logging.info(f"{self.name()}, {df.columns.tolist()}")
         df.replace({"species": {"Homo Sapiens": "Homo sapiens", "Mus Musculus": "Mus musculus"}}, inplace=True)
 
         df = self.filter_values(df, filters)
@@ -349,7 +350,7 @@ class LncReg(Interactions):
 
     def load_network(self, file_resources, source_col_name, target_col_name, edge_attr, directed, filters):
[…]
