### 2021-01-01 40c1df9
Pin large-image to latest version 1.3.2
