+++ requirements.txt
@@ -1,5 +1,5 @@
 biopython
-large-image
+large-image==1.3.2
 bokeh
 umap-learn
 dask[complete]
