### 2021-01-01 b0c772f
Pin bokeh to latest version 2.2.3
