### 2021-01-01 e1a01bc
Pin umap-learn to latest version 0.4.6
