+++ requirements.txt
@@ -1,7 +1,7 @@
 biopython
 large-image
 bokeh
-umap-learn
+umap-learn==0.4.6
 dask[complete]
 networkx>=2.1
 numpy>=1.16.5
