### 2021-01-01 69f13d8
Pin matplotlib to latest version 3.3.3
