### 2021-01-01 776a957
Pin scipy to latest version 1.6.0
