### 2021-01-01 ffd6f3a
Pin astropy to latest version 4.2
