+++ requirements.txt
@@ -8,7 +8,7 @@ numpy>=1.16.5
 pandas
 matplotlib
 scipy
-astropy
+astropy==4.2
 h5py
 requests
 setuptools
