### 2021-01-01 2fc1b74
Pin h5py to latest version 3.1.0
