### 2021-01-01 815e3ba
Pin typing to latest version 3.7.4.3
