### 2021-01-01 e2000ea
Pin bioservices to latest version 1.7.10
