### 2021-01-01 06f61eb
Pin itsdangerous to latest version 1.1.0
