### 2021-01-01 ef5a1b9
Pin dash to latest version 1.18.1
