+++ requirements.txt
@@ -16,7 +16,7 @@ typing
 bioservices
 itsdangerous
 dash
-dash-daq
+dash-daq==0.5.0
 dash-bootstrap-components
 gunicorn
 six
