### 2021-01-01 6954721
Pin six to latest version 1.15.0
