### 2021-01-01 c116e1a
Pin scikit-allel to latest version 1.3.2
