### 2021-01-01 a8828f0
Pin rarfile to latest version 4.0
