### 2021-01-01 3518bac
Pin goatools to latest version 1.0.15
