### 2021-01-01 57c1e38
Pin obonet to latest version 0.2.6
