### 2021-01-01 #30 @JonnyTran
> @NickleDave
> Please, whenever you are ready, go ahead and open a new issue that's a full submission review.
> If you can think of reviewers that would be appropriate, please also feel free to suggest them in the initial comment.
> I will also look for reviewers with relevant expertise when I start the process the week of Jan. 11.
> I can close this issue once you create the issue for full review.

I've created a new issue for the **full review** at https://github.com/pyOpenSci/software-review/issues/31. I haven't think of any reviewers yet, but I'll update the issue if I find any.
