### 2021-01-09 8c725a9
Pin setuptools to latest version 51.1.2
