+++ docs/usage.rst
@@ -2,20 +2,18 @@
 Usage
 =====
 
-To use openOmics in a project::
-::
-
-   import openomics
-
-Importing the openomics library & modules
------------------------------------------
+To use openomics in a project
+-----------------------------
 
 .. code:: python
 
-   from openomics import MultiOmics, MessengerRNA, MicroRNA, LncRNA
+   import openomics
+
 
 Import GTEx Tissue-specific gene expression dataset (directly from URL)
 -----------------------------------------------------------------------
+We start with tissue-specific gene expression data set from The Genotype-Tissue Expression (GTEx) project. First, we load the GTEx database from the path https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/, which automatically download and parse the gene_median_tpm files to create gene expression matrix. The matrix can be accessed at `gtex.data`.
+
 
 .. code:: python
 
@@ -28,12 +26,19 @@ Import GTEx Tissue-specific gene expression dataset (directly from URL)
    gene_id = pd.Index(gtex_transcripts["gene_id"].unique())
    gene_name = pd.Index(gtex_transcripts["gene_name"].unique())
 
+   gtex_transcripts.head()
+
 Import GENCODE human release 32
 -------------------------------
+Next, we can annotate the genes in our GTEx expression dataset with genomics annotation from GENCODE. In this example, we use the URL path prefix "ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/" which specifies the species and release version. We also pass a dictionary `file_resources`, with key-value pairs where the key is name of file and value is the suffix of the file download URL.
+
+For example, file_resources={"long_noncoding_RNAs.gtf": "gencode.v32.long_noncoding_RNAs.gtf.gz"} will download file located at `ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/gencode.v32.long_noncoding_RNAs.gtf.gz` to process the `long_noncoding_RNAs.gtf` file.
+
+Here, we loads both "long_noncoding_RNAs.gtf" and "basic.annotation.gtf" which builds a dataframe of combined annotations for both lncRNAs and mRNAs. You can specify different annotation files options from GENCODE by modifying the `file_resources` dict argument.
 
 .. code:: python
 
-   from openomics.database import GENCODE
+   from openomics.database import GENCODE, EnsemblGenes
 
    gencode = GENCODE(path="ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_32/",
                      file_resources={"long_noncoding_RNAs.gtf": "gencode.v32.long_noncoding_RNAs.gtf.gz",
@@ -42,32 +47,41 @@ Import GENCODE human release 32
                                      "transcripts.fa": "gencode.v32.transcripts.fa.gz"},
                      remove_version_num=True)
 
+   # We also loads Ensembl genes to get list of miRNA gene IDs
+   ensembl = EnsemblGenes(biomart='hsapiens_gene_ensembl', npartitions=8, )
+
 Load single omics expressions for MessengerRNA, MicroRNA, LncRNA
 ----------------------------------------------------------------
+We instantiate the MessengerRNA, MicroRNA and LncRNA -omics expression data from `gtex.data`. Since the gene expression were not seperated by RNA type, we use GENCODE and Ensembl gene annotations to filter the list of mRNA, miRNA, and lncRNAs.
 
 .. code:: python
 
[…]
