+++ docs/installation.rst
@@ -32,20 +32,14 @@ You can either clone the public repository:
 
 .. code-block:: console
 
-    $ git clone git://github.com/JonnyTran/OpenOmics
-
-Or download the `tarball`_:
-
-.. code-block:: console
-
-    $ curl  -OL https://github.com/JonnyTran/OpenOmics/tarball/master
+    $ git clone https://github.com/JonnyTran/OpenOmics
 
 Once you have a copy of the source, you can install it with:
 
 .. code-block:: console
 
-    $ python setup.py install
+    $ pip install .openomics/ -e
 
 
-.. _Github repo: https://github.com/JonnyTran/OpenOmics
-.. _tarball: https://github.com/JonnyTran/OpenOmics/tarball/master
+.. _Github repo: https://github.com/BioMeCIS-Lab/OpenOmics
+.. _tarball: https://github.com/JonnyTran/BioMeCIS-Lab/tarball/master
