### 2021-01-14 0a62698
Pin umap-learn to latest version 0.5.0
