+++ requirements.txt
@@ -1,7 +1,7 @@
 biopython
 large-image
 bokeh
-umap-learn
+umap-learn==0.5.0
 dask[complete]
 networkx>=2.1
 numpy>=1.16.5
