### 2021-01-14 576f5e5
Pin bioservices to latest version 1.7.11
