+++ .deepsource.toml
@@ -1,0 +1,20 @@
+version = 1
+
+test_patterns = ["tests/"]
+
+exclude_patterns = ["tests/data/"]
+
+[[analyzers]]
+name = "python"
+enabled = true
+
+  [analyzers.meta]
+  runtime_version = "3.x.x"
+
+[[transformers]]
+name = "yapf"
+enabled = true
+
+[[transformers]]
+name = "autopep8"
+enabled = true
