### 2021-01-22 37b4c62
Format code with yapf and autopep8
This commit fixes the style issues introduced in 447bb6d according to the output
from yapf and autopep8.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/02a1e67c-ff88-4b03-bb74-71f441205bd2/
