+++ openomics/utils/io.py
@@ -27,8 +27,11 @@ def get_pkg_data_filename(dataurl, file, verbose):
         dataurl = dataurl + "/"
     logging.info(f"Fetching file from: {dataurl}{file}") if verbose else None
 
-    with data.conf.set_temp("dataurl", dataurl), data.conf.set_temp("remote_timeout", 30):
-        return data.get_pkg_data_filename(file, package="openomics.database", show_progress=True)
+    with data.conf.set_temp("dataurl",
+                            dataurl), data.conf.set_temp("remote_timeout", 30):
+        return data.get_pkg_data_filename(file,
+                                          package="openomics.database",
+                                          show_progress=True)
 
 
 def read_db(path, table, index_col):
@@ -45,7 +48,10 @@ def read_db(path, table, index_col):
     uid, dt = list(table.columns)
     q = sa.select([dt.cast(sa.types.String)]).select_from(table)
 
-    daskDF = dd.read_sql_table(table, path, index_col=index_col, parse_dates={'datetime': '%Y-%m-%d %H:%M:%S'})
+    daskDF = dd.read_sql_table(table,
+                               path,
+                               index_col=index_col,
+                               parse_dates={'datetime': '%Y-%m-%d %H:%M:%S'})
     return daskDF
 
 
@@ -68,7 +74,8 @@ def retry(num=5):
 
     """
     s = requests.Session()
-    retries = Retry(total=num, backoff_factor=0.1,
+    retries = Retry(total=num,
+                    backoff_factor=0.1,
                     status_forcelist=[500, 502, 503, 504])
     s.mount('http://', HTTPAdapter(max_retries=retries))
 
