+++ openomics/__init__.py
@@ -1,4 +1,16 @@
 from __future__ import print_function, division, absolute_import
+from .visualization import (umap)
+from .multiomics import (MultiOmics)
+from .clinical import ClinicalData
+from .proteomics import (Protein)
+from .genomics import (SomaticMutation, DNAMethylation, CopyNumberVariation)
+from .transcriptomics import (
+    ExpressionData,
+    MessengerRNA,
+    MicroRNA,
+    LncRNA,
+)
+from . import database, utils
 
 import imp
 import sys
@@ -7,7 +19,6 @@ import dask.dataframe as dd
 import pandas as pd
 
 # -*- coding: utf-8 -*-
-
 """Top-level package for openomics."""
 
 __author__ = """Nhat (Jonny) Tran"""
@@ -16,30 +27,6 @@ __version__ = '0.8.4'
 
 __BACKEND__ = "pandas"
 backend = pd
-
-from . import database, utils
-
-from .transcriptomics import (
-    ExpressionData, MessengerRNA, MicroRNA, LncRNA,
-)
-
-from .genomics import (
-    SomaticMutation, DNAMethylation, CopyNumberVariation
-)
-
-from .proteomics import (
-    Protein
-)
-
-from .clinical import ClinicalData
-
-from .multiomics import (
-    MultiOmics
-)
-
-from .visualization import (
-    umap
-)
 
 
 def set_backend(new):

[…]
