### 2021-01-22 5ac7d26
Use identity check for comparison with True
