### 2021-01-22 adea676
change test_patterns
