+++ .deepsource.toml
@@ -1,6 +1,6 @@
 version = 1
 
-test_patterns = ["tests/"]
+test_patterns = ["tests/test_*"]
 
 exclude_patterns = ["tests/data/"]
 
