### 2021-01-22 d8c6382
Format code with yapf and autopep8
This commit fixes the style issues introduced in adea676 according to the output
from yapf and autopep8.

Details: https://deepsource.io/gh/BioMeCIS-Lab/OpenOmics/transform/ad395550-cf01-408e-9800-ca94b868d3c8/
