+++ tests/test_annotations.py
@@ -4,7 +4,8 @@ from .test_multiomics import *
 
 @pytest.fixture
 def generate_RNACentral_ftp():
-    return RNAcentral(path="ftp://ftp.ebi.ac.uk/pub/databases/RNAcentral/current_release/")
+    return RNAcentral(
+        path="ftp://ftp.ebi.ac.uk/pub/databases/RNAcentral/current_release/")
 
 
 def test_import_rnacentral_db(generate_RNACentral_ftp):
@@ -21,16 +22,18 @@ def test_rnacentral_annotate(generate_TCGA_LUAD, generate_RNACentral_ftp):
         generate_TCGA_LUAD:
         generate_RNACentral_ftp:
     """
-    generate_TCGA_LUAD.MessengerRNA.annotate_genomics(database=generate_RNACentral_ftp, index="gene_name",
-                                                      columns=['gene_name', 'transcript_id', 'RNA type', 'go_id',
-                                                               'Rfams'])
+    generate_TCGA_LUAD.MessengerRNA.annotate_genomics(
+        database=generate_RNACentral_ftp,
+        index="gene_name",
+        columns=['gene_name', 'transcript_id', 'RNA type', 'go_id', 'Rfams'])
     assert {'transcript_id', 'RNA type', 'go_id', 'Rfams'}.issubset(
         generate_TCGA_LUAD.MessengerRNA.get_annotations().columns)
 
 
 @pytest.fixture
 def generate_GTEx_expressions():
-    return GTEx(path="https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/", )
+    return GTEx(
+        path="https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/", )
 
 
 def test_import_GTEx(generate_GTEx_expressions):
@@ -39,7 +42,8 @@ def test_import_GTEx(generate_GTEx_expressions):
         generate_GTEx_expressions:
     """
     assert generate_GTEx_expressions.data_path == "https://storage.googleapis.com/gtex_analysis_v8/rna_seq_data/"
-    assert not generate_GTEx_expressions.get_expressions(index="gene_name").empty
+    assert not generate_GTEx_expressions.get_expressions(
+        index="gene_name").empty
     assert not generate_GTEx_expressions.get_expressions(index="gene_id").empty
 
 
@@ -49,10 +53,14 @@ def test_GTEx_annotate(generate_TCGA_LUAD, generate_GTEx_expressions):
         generate_TCGA_LUAD:
         generate_GTEx_expressions:
     """
-    generate_TCGA_LUAD.MessengerRNA.annotate_expressions(database=generate_GTEx_expressions, index="gene_name")
-    generate_TCGA_LUAD.LncRNA.annotate_expressions(database=generate_GTEx_expressions, index="gene_id")
-    generate_TCGA_LUAD.MicroRNA.annotate_expressions(database=generate_GTEx_expressions, index="gene_name")
-    assert not generate_TCGA_LUAD.MessengerRNA.get_annotation_expressions().empty
+    generate_TCGA_LUAD.MessengerRNA.annotate_expressions(
+        database=generate_GTEx_expressions, index="gene_name")
+    generate_TCGA_LUAD.LncRNA.annotate_expressions(
+        database=generate_GTEx_expressions, index="gene_id")
+    generate_TCGA_LUAD.MicroRNA.annotate_expressions(
+        database=generate_GTEx_expressions, index="gene_name")
+    assert not generate_TCGA_LUAD.MessengerRNA.get_annotation_expressions(
+    ).empty
     assert not generate_TCGA_LUAD.LncRNA.get_annotation_expressions().empty
[…]
