### 2021-01-23 59fa728
added paper.bib and started paper.md for JOSS submission
