+++ inst/paper.bib
@@ -1,0 +1,483 @@
+@article{jalili2020galaxy,
+    title = {The Galaxy platform for accessible, reproducible and collaborative biomedical analyses: 2020 update},
+    author = {Jalili, Vahid and Afgan, Enis and Gu, Qiang and Clements, Dave and Blankenberg, Daniel and Goecks, Jeremy and Taylor, James and Nekrutenko, Anton},
+    journal = {Nucleic Acids Research},
+    year = {2020}
+}
+
+@techreport{hagberg2008exploring,
+    title = {Exploring network structure, dynamics, and function using NetworkX},
+    author = {Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
+    year = {2008},
+    institution = {Los Alamos National Lab.(LANL), Los Alamos, NM (United States)}
+}
+
+@article{zhang2014proteogenomic,
+    title = {Proteogenomic characterization of human colon and rectal cancer},
+    author = {Zhang, Bing and Wang, Jing and Wang, Xiaojing and Zhu, Jing and Liu, Qi and Shi, Zhiao and Chambers, Matthew C and Zimmerman, Lisa J and Shaddox, Kent F and Kim, Sangtae and others},
+    journal = {Nature},
+    volume = {513},
+    number = {7518},
+    pages = {382--387},
+    year = {2014},
+    publisher = {Nature Publishing Group}
+}
+
+@article{ren2016integration,
+    title = {Integration of metabolomics and transcriptomics reveals major metabolic pathways and potential biomarker involved in prostate cancer},
+    author = {Ren, Shancheng and Shao, Yaping and Zhao, Xinjie and Hong, Christopher S and Wang, Fubo and Lu, Xin and Li, Jia and Ye, Guozhu and Yan, Min and Zhuang, Zhengping and others},
+    journal = {Molecular \& Cellular Proteomics},
+    volume = {15},
+    number = {1},
+    pages = {154--163},
+    year = {2016},
+    publisher = {ASBMB}
+}
+
+@article{rohart2017mixomics,
+    title = {mixOmics: An R package for ‘omics feature selection and multiple data integration},
+    author = {Rohart, Florian and Gautier, Beno{\^\i}t and Singh, Amrit and L{\^e} Cao, Kim-Anh},
+    journal = {PLoS computational biology},
+    volume = {13},
+    number = {11},
+    pages = {e1005752},
+    year = {2017},
+    publisher = {Public Library of Science}
+}
+
+@article{boekel2015multi,
+    title = {Multi-omic data analysis using Galaxy},
+    author = {Boekel, Jorrit and Chilton, John M and Cooke, Ira R and Horvatovich, Peter L and Jagtap, Pratik D and K{\"a}ll, Lukas and Lehti{\"o}, Janne and Lukasse, Pieter and Moerland, Perry D and Griffin, Timothy J},
+    journal = {Nature biotechnology},
+    volume = {33},
+    number = {2},
+    pages = {137--139},
+    year = {2015},
+    publisher = {Nature Publishing Group}
+}
+
+@article{sangaralingam2019multi,
+    title = {‘Multi-omic’data analysis using O-miner},
+    author = {Sangaralingam, Ajanthah and Dayem Ullah, Abu Z and Marzec, Jacek and Gadaleta, Emanuela and Nagano, Ai and Ross-Adams, Helen and Wang, Jun and Lemoine, Nicholas R and Chelala, Claude},
+    journal = {Briefings in bioinformatics},
+    volume = {20},
+    number = {1},
+    pages = {130--143},
+    year = {2019},
+    publisher = {Oxford University Press}
+}
+
+@article{ovaska2010large,
[…]
