### 2021-01-23 ffe0c04
added paper.bib and started paper.md for JOSS submission
