+++ inst/paper.md
@@ -1,114 +1,211 @@
-```markdown
 ---
-title: 'OpenOmics: Bioinformatics Library for Integrating Multi-omics, Annotation, and Interaction Data'
-tags:
-  - Python
-  - astronomy
-  - dynamics
-  - galactic dynamics
-  - milky way
-authors:
-  - name: Adrian M. Price-Whelan^[Custom footnotes for e.g. denoting who the corresponding author is can be included like this.]
-    orcid: 0000-0003-0872-7098
-    affiliation: "1, 2" # (Multiple affiliations must be quoted)
-  - name: Author Without ORCID
-    affiliation: 2
-  - name: Author with no affiliation
-    affiliation: 3
-affiliations:
- - name: Lyman Spitzer, Jr. Fellow, Princeton University
-   index: 1
- - name: Institution Name
-   index: 2
- - name: Independent Researcher
-   index: 3
-date: 13 August 2017
-bibliography: paper.bib
+title: 'OpenOmics: Bioinformatics Library to Integrate Multi-omics and Annotation Data' tags:
+
+- Python
+- bioinformatics
+- multiomics
+- data integration
+- big data authors:
+- name: Nhat C. Tran^[corresponding author]
+  orcid: 0000-0002-2575-9633 affiliation: 1
+- name: Jean X. Gao affiliation: 1 affiliations:
+- name: Department of Computer Science and Engineering, The University of Texas at Arlington index: 1 date: 25 January
+  2021 bibliography: paper.bib
 
 # Optional fields if submitting to a AAS journal too, see this blog post:
+
 # https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
-aas-journal: Astrophysical Journal <- The name of the AAS journal.
+
+# aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
+
+# aas-journal: Astrophysical Journal <- The name of the AAS journal.
 ---
 
 # Summary
 
-The forces on stars, galaxies, and dark matter under external gravitational
-fields lead to the dynamical evolution of structures in the universe. The orbits
-of these bodies are therefore key to understanding the formation, history, and
-future state of galaxies. The field of "galactic dynamics," which aims to model
-the gravitating components of galaxies to study their structure and evolution,
-is now well-established, commonly taught, and frequently used in astronomy.
-Aside from toy problems and demonstrations, the majority of problems require
-efficient numerical tools, many of which require the same base code (e.g., for
-performing numerical orbit integration).
+Leveraging large-scale multi-omics data is emerging as the primary approach for systemic research of human diseases and
+general biological processes. As data integration and feature engineering are the vital steps in these bioinformatics
+projects, there currently lacks a tool for standardized preprocessing of heterogeneous multi-omics and annotation data
+within the context of a clinical cohort. OpenOmics is a Python library for integrating heterogeneous multi-omics data
+and interfacing with popular public annotation databases, e.g., GENCODE, Ensembl, BioGRID. The library is designed to be
+highly flexible to allow the user to parameterize the construction of integrated datasets, interactive to assist complex
+data exploratory analyses, and scalable to facilitate working with large datasets on standard machines. We demonstrate
+design software choices to support the wide-ranging use cases of OpenOmics with the goal of maximizing usability and
+reproducibility of the data integration framework.
 
 # Statement of need
 
-`Gala` is an Astropy-affiliated Python package for galactic dynamics. Python
-enables wrapping low-level languages (e.g., C) for speed without losing
-flexibility or ease-of-use in the user-interface. The API for `Gala` was
-designed to provide a class-based and user-friendly interface to fast (C or
[…]
