### 2021-01-23 f84b101
added paper.bib and started paper.md for JOSS submission
