+++ inst/paper.md
@@ -11,14 +11,6 @@ title: 'OpenOmics: Bioinformatics Library to Integrate Multi-omics and Annotatio
 - name: Jean X. Gao affiliation: 1 affiliations:
 - name: Department of Computer Science and Engineering, The University of Texas at Arlington index: 1 date: 25 January
   2021 bibliography: paper.bib
-
-# Optional fields if submitting to a AAS journal too, see this blog post:
-
-# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-
-# aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
-
-# aas-journal: Astrophysical Journal <- The name of the AAS journal.
 ---
 
 # Summary
