### 2021-01-23 fff5a7f
added paper.bib and started paper.md for JOSS submission
