### 2021-01-23 8294d05
added paper.bib and started paper.md for JOSS submission
