+++ inst/paper.bib
@@ -12,6 +12,60 @@
     institution = {Los Alamos National Lab.(LANL), Los Alamos, NM (United States)}
 }
 
+@article{cancer2014comprehensive,
+  title={Comprehensive molecular profiling of lung adenocarcinoma},
+  author={Cancer Genome Atlas Research Network and others},
+  journal={Nature},
+  volume={511},
+  number={7511},
+  pages={543--550},
+  year={2014},
+  publisher={Nature Publishing Group}
+}
+
+@article{hassan2020integration,
+  title={Integration of Transcriptome and Metabolome Provides Unique Insights to Pathways Associated With Obese Breast Cancer Patients},
+  author={Hassan, Mohammed A and Al-Sakkaf, Kaltoom and Shait Mohammed, Mohammed Razeeth and Dallol, Ashraf and Al-Maghrabi, Jaudah and Aldahlawi, Alia and Ashoor, Sawsan and Maamra, Mabrouka and Ragoussis, Jiannis and Wu, Wei and others},
+  journal={Frontiers in Oncology},
+  volume={10},
+  pages={804},
+  year={2020},
+  publisher={Frontiers}
+}
+
+@article{stuart2019integrative,
+  title={Integrative single-cell analysis},
+  author={Stuart, Tim and Satija, Rahul},
+  journal={Nature Reviews Genetics},
+  volume={20},
+  number={5},
+  pages={257--272},
+  year={2019},
+  publisher={Nature Publishing Group}
+}
+
+@article{cervera2019anduril,
+  title={Anduril 2: upgraded large-scale data integration framework},
+  author={Cervera, Alejandra and Rantanen, Ville and Ovaska, Kristian and Laakso, Marko and Nu{\~n}ez-Fontarnau, Javier and Alkodsi, Amjad and Casado, Julia and Facciotto, Chiara and H{\"a}kkinen, Antti and Louhimo, Riku and others},
+  journal={Bioinformatics},
+  volume={35},
+  number={19},
+  pages={3815--3817},
+  year={2019},
+  publisher={Oxford University Press}
+}
+
+@article{lappalainen2013transcriptome,
+  title={Transcriptome and genome sequencing uncovers functional variation in humans},
+  author={Lappalainen, Tuuli and Sammeth, Michael and Friedl{\"a}nder, Marc R and Ac‘t Hoen, Peter and Monlong, Jean and Rivas, Manuel A and Gonzalez-Porta, Mar and Kurbatova, Natalja and Griebel, Thasso and Ferreira, Pedro G and others},
+  journal={Nature},
+  volume={501},
+  number={7468},
+  pages={506--511},
+  year={2013},
+  publisher={Nature Publishing Group}
+}
+
 @article{zhang2014proteogenomic,
     title = {Proteogenomic characterization of human colon and rectal cancer},
     author = {Zhang, Bing and Wang, Jing and Wang, Xiaojing and Zhu, Jing and Liu, Qi and Shi, Zhiao and Chambers, Matthew C and Zimmerman, Lisa J and Shaddox, Kent F and Kim, Sangtae and others},
