### 2021-01-23 b112252
added paper.bib and started paper.md for JOSS submission
