### 2021-01-23 4861e5a
added paper.bib and started paper.md for JOSS submission
